chore: Rename pallas-machines as pallas-miniprotocols
