+++ pallas-miniprotocols/examples/blockfetch.rs
@@ -1,13 +1,13 @@
 use net2::TcpStreamExt;
-use pallas_machines::primitives::Point;
+use pallas_miniprotocols::Point;
 use std::net::TcpStream;
 
-use pallas_machines::blockfetch::{BatchClient, NoopObserver};
-use pallas_machines::handshake::{
+use pallas_miniprotocols::blockfetch::{BatchClient, NoopObserver};
+use pallas_miniprotocols::handshake::{
     n2n::{Client, VersionTable},
     MAINNET_MAGIC,
 };
-use pallas_machines::machines::run_agent;
+use pallas_miniprotocols::run_agent;
 use pallas_multiplexer::Multiplexer;
 
 fn main() {
