+++ pallas-miniprotocols/examples/chainsync-blocks.rs
@@ -1,11 +1,11 @@
 use pallas_alonzo::{crypto, Block, BlockWrapper, Fragment};
 
-use pallas_machines::chainsync::{BlockLike, Consumer, NoopObserver};
-use pallas_machines::handshake::n2c::{Client, VersionTable};
-use pallas_machines::handshake::MAINNET_MAGIC;
-use pallas_machines::machines::run_agent;
-use pallas_machines::machines::{DecodePayload, EncodePayload, PayloadDecoder, PayloadEncoder};
-use pallas_machines::primitives::Point;
+use pallas_miniprotocols::chainsync::{BlockLike, Consumer, NoopObserver};
+use pallas_miniprotocols::handshake::n2c::{Client, VersionTable};
+use pallas_miniprotocols::handshake::MAINNET_MAGIC;
+use pallas_miniprotocols::run_agent;
+use pallas_miniprotocols::Point;
+use pallas_miniprotocols::{DecodePayload, EncodePayload, PayloadDecoder, PayloadEncoder};
 use pallas_multiplexer::Multiplexer;
 use std::os::unix::net::UnixStream;
