+++ pallas-miniprotocols/examples/handshake-client.rs
@@ -1,10 +1,9 @@
 use net2::TcpStreamExt;
 use std::net::TcpStream;
 
-use pallas_machines::handshake::n2c::{Client, VersionTable};
-use pallas_machines::handshake::MAINNET_MAGIC;
-use pallas_machines::machines::run_agent;
-use pallas_machines::txsubmission::NaiveProvider;
+use pallas_miniprotocols::handshake::n2c::{Client, VersionTable};
+use pallas_miniprotocols::handshake::MAINNET_MAGIC;
+use pallas_miniprotocols::run_agent;
 use pallas_multiplexer::Multiplexer;
 
 fn main() {
@@ -16,16 +15,11 @@ fn main() {
     bearer.set_nodelay(true).unwrap();
     bearer.set_keepalive_ms(Some(30_000u32)).unwrap();
 
-    let mut muxer = Multiplexer::setup(bearer, &vec![0, 4]).unwrap();
+    let mut muxer = Multiplexer::setup(bearer, &vec![0]).unwrap();
 
     let mut hs_channel = muxer.use_channel(0);
     let versions = VersionTable::v1_and_above(MAINNET_MAGIC);
     let last = run_agent(Client::initial(versions), &mut hs_channel).unwrap();
+
     println!("{:?}", last);
-
-    let mut ts_channel = muxer.use_channel(4);
-    let ts = NaiveProvider::initial(vec![]);
-    let ts = run_agent(ts, &mut ts_channel).unwrap();
-
-    println!("{:?}", ts);
 }
