+++ pallas-miniprotocols/src/lib.rs
@@ -1,10 +1,15 @@
 mod codec;
-pub mod machines;
+mod machines;
 mod payloads;
-pub mod primitives;
+mod primitives;
 
 pub mod blockfetch;
 pub mod chainsync;
 pub mod handshake;
 pub mod localstate;
 pub mod txsubmission;
+
+pub use codec::*;
+pub use machines::*;
+pub use payloads::*;
+pub use primitives::*;
