+++ pallas/Cargo.toml
@@ -14,7 +14,7 @@ authors = [
 
 [dependencies]
 pallas-multiplexer = { version = "0.4.0", path = "../pallas-multiplexer/" }
-pallas-machines = { version = "0.4.0", path = "../pallas-machines/" }
+pallas-miniprotocols = { version = "0.4.0", path = "../pallas-miniprotocols/" }
 pallas-byron = { version = "0.4.0", path = "../pallas-byron/" }
 pallas-alonzo = { version = "0.4.0", path = "../pallas-alonzo/" }
 pallas-crypto = { version = "0.4.0", path = "../pallas-crypto/" }
